+++ src/api/core/config.py
@@ -15,7 +15,6 @@ class Config(BaseSettings):
     LANGSMITH_API_KEY: str
     LANGSMITH_PROJECT: str
     RAG_PROMPT_TEMPLATE_PATH: str = "src/api/rag/prompts/rag_generation.yaml"
-    POSTGRES_CONN_STRING: str = "postgresql://langgraph_user:langgraph_password@postgres:5432/langgraph_db"
 
     model_config = SettingsConfigDict(env_file=".env")
