+++ src/chatbot_ui/streamlit_app.py
@@ -1,6 +1,5 @@
 import streamlit as st
 import requests
-import uuid
 
 from src.chatbot_ui.core.config import settings
 
@@ -9,13 +8,6 @@ st.set_page_config(
     layout="wide",
     initial_sidebar_state="expanded",
 )
-
-def get_session_id():
-    if 'session_id' not in st.session_state:
-        st.session_state.session_id = str(uuid.uuid4())
-    return st.session_state.session_id
-
-session_id = get_session_id()
 
 
 def api_call(method, url, **kwargs):
@@ -100,7 +92,7 @@ if prompt := st.chat_input("Hello! How can I assist you today?"):
         st.markdown(prompt)
     
     with st.spinner("Thinking..."):
-        status, output = api_call("post", f"{settings.API_URL}/rag", json={"query": prompt, "thread_id": session_id})
+        status, output = api_call("post", f"{settings.API_URL}/rag", json={"query": prompt})
         # Update retrieved items
         st.session_state.retrieved_items = output.get("used_image_urls", [])
